board; cell; box types and functions split into respective files; magic numbers moved to consts/vars; support for other board sizes (2x2, 4x4)
